extractPins()
At first, I did not want to implement the pin extraction.
But when I began to work on the net extraction, some pin-to-gate nets made it clear that the pin cannot be left aside.
Indeed, those nets still need to be routed and can be important in the partitioning.

However, I may need to keep to pins pinned in place and not allow the partitioner to move them around. Not sure yet.