New script to draw an image of the design.
It uses the classes and functions of def_parser.
The image size is the die size in tenth of micrometers, that is a tenth of the unit base used in the Design class.
For each gate of the design, its origin is added to the image as a white pixel.