New clustering method: progressive wirelength.
The basic idea is to cluster to shortest nets together.
To do that, we sort them based on their length and then pick them one at a time, check if they connected different clusters and if so, merge those. We keep doing this until the cluster amount objective is reached.

The aim is to hide the shortest nets and only work on the longest ones in the partitioning pahse.
Now we need to see what happens when we try to minimize the number of nets cut or maximize the total length cut.

Also, now the Design.clusters is a dictionary.