def_parser: new clustering: hierarchical geometric.
This follows the same idea as the naive geometric.
However, when the former splits the design in subparts with the same aspect ratio as the global design, the later splits recursively the design into smaller parts.

It always first cut verticaly then horizontaly.
This scheme ensures that different clustering grains are correctly comparable.