Update cluster analysis: no more matrix.
It was taking a long time and a lot of memory, as the matrix was stored as an array of arrays, the size of the number of clusters.

Timing: (1) without any analysis: 28.8s, (2) with a complete analysis: 43.5s, (3) with a limited analysis (no matrices): 28.6s.

Getting rid of that matric analysis is no big deal, I only used it for a short while a few years back.