DEF parser: Pin handling problems.
1. Some PORTs in the LEF are defined using a POLYGON statement. Those were not handled up to now, ignoring those PORTs entirely.
Now, I find an overlaping rectangle on top of the polygon to approximate it.
I just want an origin, width and height, nothing more to keep the code simple.

2. Sometimes, the PnR tool does not place the PINs. In that case, I assigned them default coordinates: (0,0).
However, when a PIN is not placed, it is not taken into account for the routing either.
The solution is thus twofold:
a) Find a legitimate placement for the PIN.
I list all the PORTs it might be connected to and find the closest to an edge.
That PORT becomes the most probable connection point for the pin. I then project the coordinates to the closest edge of the die.
That becomes the coordinates of the PIN.

b) Evaluate the additional wirelength
As the PIN is not taken into account in the routing, we miss a bit of the wirelength or even the whole of it when the net only connects a PIN and a single standard cell.
This additional WL is simply the distance between the PORT identified earlier and the PIN.

Another fix in this commit: the Bounding boxes where all wrong as I did not consider the absolute coordinates of the PORTs!